backend/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
flask_cors
flask_jwt_extended